requirements.txt
# Healix — Python dependencies
#
# Install with:
#   pip install -r requirements.txt

# Google Agent Development Kit + Gemini SDK
google-adk>=1.0.0
google-genai>=1.0.0

# HTTP client for health API calls
httpx>=0.27.0

# Data validation (used internally by ADK)
pydantic>=2.0.0

# Environment variable loading
python-dotenv>=1.0.0

# Persistent session storage (SQLite via async SQLAlchemy)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
